requirements.txt
faiss-cpu
langchain
google-generativeai
PyMuPDF
python-dotenv
streamlit
tiktoken
numpy
pandas
scikit-learn